// lib/tom/types.ts
// Theory of Mind Types
// lib/tom/types.ts

export interface TheoryOfMindProfile {
  entityId: string;
  coreTraits: any;
  motivationalDrivers: any;
  stressTriggers: any;
  soothingPatterns: any;
  conflictPatterns: any;
  communicationStyle: any;
  currentStateHypothesis: any;
  summary: string;
}